refactor(datasheet): suppression de livraison mutualisée dans un hook
